pyproject.toml
[tool.poetry]
name = "paidcontentplatform"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
ipython = "^8.27.0"
django-stubs = "^5.1.0"
stripe = "^11.1.0"
smsaero-api = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
